Move tests and add global usings
