+++ test/TuDa.CIMS.Api.Test/Services/InvoiceGenerationServiceTest.cs
@@ -1,6 +1,4 @@
-﻿using Bogus;
-using JetBrains.Annotations;
-using TuDa.CIMS.Api.Interfaces;
+﻿using TuDa.CIMS.Api.Interfaces;
 using TuDa.CIMS.Api.Services;
 using TuDa.CIMS.Shared.Entities;
 using TuDa.CIMS.Shared.Test.Faker;
