+++ test/TuDa.CIMS.Api.Test/Usings.cs
@@ -1,2 +1,4 @@
-﻿global using FluentAssertions;
+﻿global using Bogus;
+global using FluentAssertions;
+global using JetBrains.Annotations;
 global using Moq;
